Added Part 2 of Assignment 1
Part 2 of Assignment 1 required that I create a map to store the number of times I had seen a certain phrase of words while reading from a text file. I also had to print out the information stored in the map. The phrase length (number of words that make up the phrase), was only one word for this part of the assignment, but will be expanded to include phrases of varying length for the next part of the assignment.