Added Part 3 of Assignment 1
Part 3 required me to print statistical information based on a given mode about the frequencies of all the phrases that could be made from the text of different files. I used maps of varying types to store the frequency information, store maps of those maps, and store the swapped keys and values of one map in another so I could have the map sorted by frequency rather than by alphabetical order of the phrases.